fix(skyulf-core): LabelEncoder pandas output silently upcast to float64 on unseen categories
_label_apply_pandas() built encoded columns via
`.astype(str).map(mapping).fillna(missing_code)`; the moment `.map()`
produced NaN for an unseen category, pandas silently upcasts the whole
column to float64, and `.fillna()` does not downcast back to int - so
missing_code (e.g. -1) ended up stored as -1.0 instead of -1, diverging
from the Polars path's explicit `.cast(pl.Int64)`.

Added explicit `.astype("int64")` after `.fillna()` in both the feature
and target (y) encoding branches to match Polars' guarantee.

Added 3 regression tests (pandas dtype on features, pandas dtype on
target, cross-engine value+dtype parity). Full suite: 2692 passed.
ruff/ty clean. No frontend changes needed - LabelEncoder has no
dedicated frontend node/config.

+++ skyulf-core/skyulf/preprocessing/encoding/label.py
@@ -95,11 +95,17 @@ def _label_apply_pandas(X: Any, y: Any, params: dict[str, Any]) -> tuple[Any, An
         for col in cols:
             if col in X_out.columns and col in encoders:
                 mapping = _le_mapping(encoders[col])
-                X_out[col] = X_out[col].astype(str).map(mapping).fillna(missing_code)
+                # `.map()` returns float64 the moment any unseen value maps to
+                # NaN (even after `.fillna(missing_code)` the column stays
+                # float, since pandas never upcasts back to int) - explicit
+                # `.astype("int64")` matches the Polars path's `.cast(pl.Int64)`.
+                X_out[col] = (
+                    X_out[col].astype(str).map(mapping).fillna(missing_code).astype("int64")
+                )
 
     if y_out is not None and "__target__" in encoders:
         mapping = _le_mapping(encoders["__target__"])
-        y_out = y_out.astype(str).map(mapping).fillna(missing_code)
+        y_out = y_out.astype(str).map(mapping).fillna(missing_code).astype("int64")
 
     return X_out, y_out
